Pinger Version 1.00
Edited Ping3 and turned it into pinger.sh. Made folders for old
versions. In pinger.sh, I made the graphics more attractive, as well as
adding more functionality and user-friendliness. Worked a lot on the
WHOIS function to get the WHOIS information from domains and save them
to files. Need to work on creating the statistics files next.

-Chandler